requirements.txt
langchain
openai
baidu_aip
pygame
duckduckgo-search